Fix error

+++ src/main.c
@@ -257,7 +257,7 @@ void task1(char *buffer) {
   char **pool = gen_addrs(POOL_LEN, buffer);
   char *base = pool[rand()%POOL_LEN];
   int threshold = 450;
-  int significant_bits = 0;
+  long long int significant_bits = 0;
   int num_conflicts;
   char **conflicts = get_conflicts(buffer, threshold, &num_conflicts);
